Additional refactoring regarding utils/torch_utils.py
