+++ src/gobi_cath_classification/scripts_finn/baseline_models.py
@@ -4,8 +4,8 @@ import numpy as np
 import pandas as pd
 import torch
 from gobi_cath_classification.pipeline.model_interface import ModelInterface, Prediction
-from gobi_cath_classification.pipeline import torch_utils
-from gobi_cath_classification.pipeline.torch_utils import set_random_seeds
+from gobi_cath_classification.pipeline.utils import torch_utils
+from gobi_cath_classification.pipeline.utils.torch_utils import set_random_seeds
 from gobi_cath_classification.pipeline.data_loading import DataSplits
 from gobi_cath_classification.pipeline.data import Dataset
 from gobi_cath_classification.pipeline.sample_weights import compute_class_counts
